update the instrumentation to Kamon 1.2 and Kanela
