+++ kamon-scala-future/src/main/resources/reference.conf
@@ -2,10 +2,17 @@
 # kamon-scala-future Reference Configuration #
 # ========================================== #
 
-kamon {
-  modules {
-    kamon-scala-future {
-      requires-aspectj = yes
-    }
+kanela.modules {
+
+  scala-future {
+    name = "Futures Intrumentation Module"
+    stoppable = false
+    instrumentations = [
+      "kamon.instrumentation.scala.FutureInstrumentation"
+    ]
+
+    within = [
+      "scala.concurrent..*"
+    ]
   }
 }
