+++ kamon-scala-future/src/main/scala-2.11/kamon/instrumentation/scala/aspectj/FutureInstrumentation.scala
@@ -14,10 +14,10 @@
  * =========================================================================================
  */
 
-package kamon.scala.instrumentation
+package kamon.instrumentation.scala.aspectj
 
 import kamon.Kamon
-import kamon.context.HasContext
+import kamon.instrumentation.Mixin.HasContext
 import org.aspectj.lang.ProceedingJoinPoint
 import org.aspectj.lang.annotation._
