+++ kamon-scala-future/src/test/scala/kamon/instrumentation/scala/FutureInstrumentationSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  * ========================================================== */
-package kamon.scala.instrumentation
+package kamon.instrumentation.scala
 
 import kamon.Kamon
 import kamon.testkit.ContextTesting
