+++ kamon-scalaz-future/src/main/resources/META-INF/aop.xml
@@ -2,7 +2,7 @@
 
 <aspectj>
   <aspects>
-    <aspect name="kamon.scalaz.instrumentation.FutureInstrumentation"/>
+    <aspect name="kamon.instrumentation.scalaz.aspectj.FutureInstrumentation"/>
   </aspects>
 
   <weaver>
