+++ kamon-scalaz-future/src/main/resources/reference.conf
@@ -1,12 +1,18 @@
-# =========================================== #
-# kamon-scalaz-future Reference Configuration #
-# =========================================== #
+# ========================================== #
+# kamon-scala-future Reference Configuration #
+# ========================================== #
 
-kamon {
+kanela.modules {
 
-  modules {
-    kamon-scalaz-future {
-      requires-aspectj = yes
-    }
+  scalaz-future {
+    name = "Scalaz Future Intrumentation Module"
+    stoppable = false
+    instrumentations = [
+      "kamon.instrumentation.scalaz.FutureInstrumentation"
+    ]
+
+    within = [
+      "scalaz.concurrent..*"
+    ]
   }
 }
