+++ kamon-twitter-future/src/main/resources/META-INF/aop.xml
@@ -2,7 +2,7 @@
 
 <aspectj>
   <aspects>
-    <aspect name="kamon.twitter.instrumentation.FutureInstrumentation"/>
+    <aspect name="kamon.twitter.instrumentation.aspectj.FutureInstrumentation"/>
   </aspects>
 
   <weaver>
