+++ kamon-twitter-future/src/main/resources/reference.conf
@@ -2,11 +2,17 @@
 # kamon-twitter-future Reference Configuration #
 # ============================================ #
 
-kamon {
+kanela.modules {
 
-  modules {
-    kamon-twitter-future {
-      requires-aspectj = yes
-    }
+  twitter-future {
+    name = "Twitter Future Intrumentation Module"
+    stoppable = false
+    instrumentations = [
+      "kamon.instrumentation.twitter.FutureInstrumentation"
+    ]
+
+    within = [
+      "com.twitter.util..*"
+    ]
   }
 }
